+++ src/tests_api/test_users.py
@@ -2,12 +2,15 @@ import pytest
 from api.main import app
 from api.pwmodels import User, Loan
 from fastapi.testclient import TestClient
+from conftest import AUTH_ADMIN, AUTH_USER
 
 client = TestClient(app)
 
 
 def test_create_user():
-    response = client.post("/users", json={"name": "bob", "email": "bob@nomail"})
+    response = client.post(
+        "/users", json={"name": "bob", "email": "bob@nomail"}, headers=AUTH_ADMIN
+    )
     assert response.status_code == 200
     newuser = response.json()
     assert "id" in newuser
@@ -18,7 +21,7 @@ def test_create_user():
     assert user_db.email == "bob@nomail"
 
     # Check in API
-    response = client.get(f"/users/{newuser['id']}")
+    response = client.get(f"/users/{newuser['id']}", headers=AUTH_ADMIN)
     assert response.status_code == 200
     user = response.json()
     assert user["name"] == "bob"
@@ -27,28 +30,32 @@ def test_create_user():
 
 def test_create_user_attributes():
     newjson = {"name": "bob", "email": "bob@nomail", "credit": 3, "role": "admin"}
-    response = client.post("/users", json=newjson)
+    response = client.post("/users", json=newjson, headers=AUTH_ADMIN)
     newUser = response.json()
 
     # Check in API
-    response = client.get(f"/users/{newUser['id']}")
+    response = client.get(f"/users/{newUser['id']}", headers=AUTH_ADMIN)
     User = response.json()
     assert newjson.items() <= User.items()
 
 
 def test_delete_user():
     "Create a user with loans, everything must be removed"
-    response = client.post("/users", json={"name": "bob", "email": "bob@nomail"})
+    response = client.post(
+        "/users", json={"name": "bob", "email": "bob@nomail"}, headers=AUTH_ADMIN
+    )
     user_id = response.json()["id"]
-    response = client.post("/items", json={"name": "obj"})
+    response = client.post("/items", json={"name": "obj"}, headers=AUTH_ADMIN)
     item_id = response.json()["id"]
     response = client.post(
-        "/loans", json={"user": user_id, "items": [item_id], "cost": 0}
+        "/loans",
+        json={"user": user_id, "items": [item_id], "cost": 0},
+        headers=AUTH_ADMIN,
     )
     loan_id = response.json()[0]["id"]
 
     # Delete via API
-    response = client.delete(f"/users/{user_id}")
+    response = client.delete(f"/users/{user_id}", headers=AUTH_ADMIN)
     assert response.status_code == 200
 
     # Check in DB
@@ -58,8 +65,16 @@ def test_delete_user():
 
 def test_delete_unknown_user():
     # Delete via API
-    response = client.delete("/users/7")
+    response = client.delete("/users/7", headers=AUTH_ADMIN)
     assert response.status_code == 404
+
+
+def test_delete_not_authenticated():
+    response = client.delete("/users/7")
+    assert response.status_code == 403
+
+    response = client.delete("/users/7", headers=AUTH_USER)
+    assert response.status_code == 403
 
 
 @pytest.mark.parametrize(
@@ -69,16 +84,16 @@ def test_delete_unknown_user():
 def test_edit_user_attributes(toedit: dict):
     # Create User
     newjson = {"name": "bob", "email": "bob@nomail", "credit": 3, "role": "admin"}
-    response = client.post("/users", json=newjson)
+    response = client.post("/users", json=newjson, headers=AUTH_ADMIN)
     newUser = response.json()
 
     # Edit with APi
     newjson |= toedit
-    response = client.post(f"/users/{newUser['id']}", json=newjson)
+    response = client.post(f"/users/{newUser['id']}", json=newjson, headers=AUTH_ADMIN)
     assert response.status_code == 200
 
     # Check in API
-    response = client.get(f"/users/{newUser['id']}")
+    response = client.get(f"/users/{newUser['id']}", headers=AUTH_ADMIN)
     User = response.json()
     assert newjson.items() <= User.items()
 
@@ -87,18 +102,18 @@ def test_get_users():
     # Create Users
     User1 = {"name": "bob", "email": "bob@nomail", "credit": 3, "role": "admin"}
     User2 = {"name": "alice", "email": "alice@nomail"}
-    response = client.post("/users", json=User1)
-    response = client.post("/users", json=User2)
+    response = client.post("/users", json=User1, headers=AUTH_ADMIN)
+    response = client.post("/users", json=User2, headers=AUTH_ADMIN)
 
     # Check in API
-    response = client.get("/users")
+    response = client.get("/users", headers=AUTH_ADMIN)
     users = response.json()
     assert len(users) == 2
     assert User1.items() <= users[0].items()
     assert User2.items() <= users[1].items()
 
     # Limit get Users to 1 User
-    response = client.get("/users?nb=1")
+    response = client.get("/users?nb=1", headers=AUTH_ADMIN)
     users = response.json()
     assert len(users) == 1
     assert User1.items() <= users[0].items()
@@ -106,23 +121,27 @@ def test_get_users():
 
 def test_get_users_loancount():
     # Create a user without loan
-    response = client.post("/users", json={"name": "bob", "email": "bob@nomail"})
+    response = client.post(
+        "/users", json={"name": "bob", "email": "bob@nomail"}, headers=AUTH_ADMIN
+    )
     user_id = response.json()["id"]
 
     # Check API
-    response = client.get("/users")
+    response = client.get("/users", headers=AUTH_ADMIN)
     user = response.json()[0]
     assert user["loans"] == 0
 
     # Create a loan
-    response = client.post("/items", json={"name": "obj"})
+    response = client.post("/items", json={"name": "obj"}, headers=AUTH_ADMIN)
     item_id = response.json()["id"]
     response = client.post(
-        "/loans", json={"user": user_id, "items": [item_id], "cost": 0}
+        "/loans",
+        json={"user": user_id, "items": [item_id], "cost": 0},
+        headers=AUTH_ADMIN,
     )
     response.json()[0]["id"]
 
     # Check API
-    response = client.get("/users")
+    response = client.get("/users", headers=AUTH_ADMIN)
     user = response.json()[0]
     assert user["loans"] == 1
